add / refactor utils (for login with gooddollar PR)
