+++ src/RouterSelector.native.js
@@ -4,7 +4,7 @@ import { first } from 'lodash'
 import Splash, { animationDuration } from './components/splash/Splash'
 import useUpdateDialog from './components/appUpdate/useUpdateDialog'
 import { delay } from './lib/utils/async'
-import retryImport from './lib/utils/retryImport'
+import { retryImport } from './lib/utils/system'
 import { handleLinks } from './lib/utils/linking'
 import { APP_OPEN, fireEvent, initAnalytics } from './lib/analytics/analytics'
