+++ src/components/signin/Mnemonics.js
@@ -13,7 +13,7 @@ import { ExceptionCategory } from '../../lib/exceptions/utils'
 import { withStyles } from '../../lib/styles'
 import { useDialog } from '../../lib/dialog/useDialog'
 import { getFirstWord } from '../../lib/utils/getFirstWord'
-import restart from '../../lib/utils/restart'
+import { restart } from '../../lib/utils/system'
 import useUserExists from '../../lib/login/useUserExists'
 import Text from '../common/view/Text'
 import Section from '../common/layout/Section'
