--- src/lib/utils/restart.native.js
@@ -1,3 +1,0 @@
-import RNRestart from 'react-native-restart'
-
-export default () => RNRestart.Restart()
