--- src/lib/utils/restart.web.js
@@ -1,10 +1,0 @@
-export default (fromUrl = null) => {
-  const { location } = window
-
-  if (!fromUrl) {
-    location.reload(true)
-    return
-  }
-
-  location.replace(fromUrl)
-}
