--- src/lib/utils/retryImport.native.js
@@ -1,5 +1,0 @@
-// @flow
-
-const retryImport = fn => fn()
-
-export default retryImport
