--- src/lib/utils/retryImport.web.js
@@ -1,6 +1,0 @@
-// @flow
-import { retry } from './async'
-
-const retryImport = fn => retry(fn, 5, 1000)
-
-export default retryImport
