+++ src/lib/utils/uri.js
@@ -1,8 +1,11 @@
 // @flow
 
+import { decode as atob, encode as btoa } from 'base-64'
+import { isString } from 'lodash'
 import isURL from 'validator/lib/isURL'
 
 import { appUrl } from '../../lib/utils/env'
+import { tryJson } from './string'
 
 const isUrlOptions = { require_tld: false, require_protocol: true }
 const createEmptyUri = () => ({ params: {}, searchParams: new URLSearchParams('') })
@@ -31,3 +34,7 @@ class CustomURL extends URL {
 export const isValidURI = (link: string = '') => isURL(link, isUrlOptions)
 
 export const createUrlObject = link => (isValidURI(link) ? new CustomURL(link) : createEmptyUri())
+
+export const decodeBase64Params = value => tryJson(atob(value))
+
+export const encodeBase64Params = value => btoa(isString(value) ? value : JSON.stringify(value))
